# internal/adapter/docsui/openapi/tl.yaml
openapi: 3.0.3
info:
  title: ANS Transparency Log API
  version: "2.0.0"
  description: |
    OpenAPI contract for the ANS Transparency Log (`ans-tl`). Paths
    match the reference TL's swagger byte-for-byte — `/v1/agents/*`,
    `/v1/log/*`, `/v1/internal/*`, C2SP tlog-tiles at root
    (`/checkpoint`, `/root-keys`, `/tile/*`), and
    `/internal/v1/producer-keys/*` for admin CRUD.

    Behavioral parity with the reference TL:

      - Event envelope is the V2 shape — nested `{payload:{logId,
        producer:{event, keyId, signature}}, schemaVersion:"V2",
        signature, status}`. V2 evolves the reference V1 with
        unified `identityCerts[]` / `serverCerts[]` arrays, typed
        `dnsRecordsProvisioned[]` records, and `SHA256:`-prefixed
        fingerprints. V0 + V1 schemas are served for historical
        verifier backwards compatibility (never emitted).
      - Receipts are SCITT COSE_Sign1 (RFC 8152 tag 18) with VDS 1
        (RFC 9162_SHA256) and the VDP inclusion proof in the
        unprotected header (label 396). See
        `internal/tl/receipt/constants.go` for the concrete labels.
      - `/root-keys` emits the sumdb-note verification format
        byte-for-byte compatible with the reference.
      - `/internal/v1/producer-keys` admin routes match the reference
        swagger.yaml snake_case DTOs; the only ans-only extension is
        `GET /internal/v1/producer-keys/{key_id}` (reference doesn't
        expose single-record lookup).
servers:
...

...
security:
  - bearerAuth: []
paths:
...
  /v1/internal/agents/event:
    post:
      tags: [Transparency Log]
      summary: Append a signed event to the log
      description: |
        Accepts a JCS-canonicalized producer-event body signed with a
        detached JWS (`X-Signature` header). Successful appends return
        the Tessera-assigned leaf index and hash; duplicate submissions
        (same SHA-256(inner event JCS)) return 200 with the original
        leaf so outbox retries are idempotent.
      operationId: appendEvent
      parameters:
        - name: X-Signature
          in: header
          required: true
          description: |
            Detached JWS compact serialization over the raw request
            body. Header shape:
            `{alg:ES256, kid:<producer-keyId>, typ:JWT, timestamp:<unix-seconds>, raid:<raId>}`.
            The `raid` value must match the `raId` in the body.
          schema:
            type: string
      requestBody:
        required: true
        description: JCS-canonicalized inner producer event (≤ 256 KiB).
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerEvent'
      responses:
        '200':
          description: |
            Event logged. On a duplicate submission (identical
            inner-event JCS hash) the response carries
            `duplicate: true` + the pre-existing leaf so retries are
            idempotent. Matches the reference TL's status-code and
            body shape.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppendResponse'
        '413':
          description: Request body exceeds 256 KiB.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: |
            Body fails validation. Codes include
            `NO_PRODUCER_SIGNATURE`, `INVALID_SIGNATURE_HEADER`,
            `NOT_FOUND_PRODUCER_KEY`, `INVALID_PRODUCER_KEY`,
            `MISMATCH_SIGNATURE`, `RAID_MISMATCH`, `INVALID_EVENT`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  # ──────────────────────────────────────────────────────────────
  # Transparency Log — verifier reads (public under public-read)
  # ──────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static API key or OIDC bearer token. Admin routes
        (`/internal/v1/producer-keys/*`) also require the caller to
        be an admin — static keys are always admin in the
        quickstart; OIDC tokens inherit admin-ness from the
        configured admin-groups list. Verifier GETs are anonymous
        when `cfg.auth.public-read: true`.
...
  schemas:
...
    ProducerEvent:
      type: object
      description: |
        The canonical inner-event shape an RA POSTs. Must be
        JCS-canonicalized when signed; the server validates the
        signature over the raw request body and re-canonicalizes.
      required: [ansId, ansName, eventType, timestamp, raId, agent]
      properties:
        ansId:      { type: string, format: uuid }
        ansName:    { type: string, example: "ans://v1.0.0.agent.example.com" }
        eventType:
          type: string
          enum: [AGENT_REGISTRATION, DOMAIN_VALIDATION, AGENT_ACTIVE, AGENT_REVOCATION]
        raId:       { type: string, example: "ans-ra-local" }
        issuedAt:   { type: string, format: date-time }
        timestamp:  { type: string, format: date-time }
        agent:
          type: object
          required: [host, name, version]
          properties:
            host:    { type: string }
            name:    { type: string }
            version: { type: string }
        attestations:
          type: object
          description: Event-type specific attestation payload.
    AppendResponse:
      type: object
      description: |
        `logId`, `message`, and `success` match the reference TL's
        AgentEventResponse byte-for-byte. `leafIndex`, `leafHashHex`,
        `duplicate`, and `treeSize` are ans-specific additions that
        expose the Merkle-tree position to callers who want to fetch a
        receipt without a second round-trip.
      required: [leafIndex, leafHashHex, duplicate, treeSize]
      properties:
        logId:       { type: string, format: uuid, description: "UUIDv7 assigned on first append; echoed on duplicate retry." }
        message:     { type: string, description: "Human-readable status message." }
        success:     { type: boolean, description: "Always true for 200 responses." }
        leafIndex:   { type: integer, format: int64 }
        leafHashHex: { type: string, description: "RFC 6962 leaf hash (hex)" }
        duplicate:   { type: boolean }
        treeSize:    { type: integer, format: int64 }

    # ── Badge + audit ─────────────────────────────────────────
...
    Problem:
      type: object
      description: RFC 7807 problem details.
      properties:
        type:   { type: string, example: "about:blank" }
        title:  { type: string }
        status: { type: integer }
        code:   { type: string, pattern: "^[A-Z_][A-Z0-9_]*$" }
        detail: { type: string }